pyproject.toml
[tool.poetry]
name = "cryptocurrency rate"
version = "0.1.0"
description = ""
authors = ["blyqq <f0ul2212@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.8.0"
asyncio = "^3.4.3"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pycoingecko = "^3.1.0"
apscheduler = "^3.10.4"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
timezonefinder = "^6.5.2"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
